requirements.txt
google-generativeai==0.8.3
langchain-google-genai==2.0.8
Flask==3.1.0
gunicorn==22.0.0
vertexai==1.71.1
python-dotenv==1.0.1